requirements.txt
fastapi
uvicorn[standard]
langchain
ollama
redis
psycopg2-binary
sqlalchemy
python-dotenv
docx
apscheduler
httpx
tqdm
